requirements.txt
# Core dependencies
pandas==2.2.0
numpy==1.26.3
pyyaml==6.0.1

# LLM integration
groq==0.4.2
python-dotenv==1.0.1

# Data handling
datasets==2.16.1
huggingface-hub==0.20.3

# Spark
pyspark==3.5.0
